refactor(team): Team 엔티티 captainId 제거 & TeamService를 Query와 Command로 분리
