+++ src/main/java/com/shootdoori/match/team/mapper/TeamMapper.java
@@ -13,7 +13,6 @@ public class TeamMapper {
 
     public Team toEntity(TeamRequestDto requestDto, Long captainId) {
         return Team.of(
-            captainId,
             requestDto.name(),
             requestDto.university(),
             parseToTeamType(requestDto.teamType()),
